refactor importgroup component
