+++ src/hooks/useFileUpload.ts
@@ -1,0 +1,103 @@
+import { useState, useRef } from "react";
+
+export const useFileUpload = () => {
+  const [selectedFile, setSelectedFile] = useState<File | null>(null);
+  const [isDragOver, setIsDragOver] = useState(false);
+  const fileInputRef = useRef<HTMLInputElement>(null);
+  const dragCounter = useRef(0);
+
+  const isValidFile = (file: File) => {
+    return (
+      file.type === "application/json" ||
+      file.name.endsWith(".md") ||
+      file.name.endsWith(".json")
+    );
+  };
+
+  const handleFileSelect = (file: File) => {
+    if (file && isValidFile(file)) {
+      setSelectedFile(file);
+    }
+  };
+
+  const handleDragOver = (e: React.DragEvent) => {
+    e.preventDefault();
+    setIsDragOver(true);
+  };
+
+  const handleDragEnter = (e: React.DragEvent) => {
+    e.preventDefault();
+    dragCounter.current += 1;
+    setIsDragOver(true);
+  };
+
+  const handleDragLeave = (e: React.DragEvent) => {
+    e.preventDefault();
+    dragCounter.current -= 1;
+    if (dragCounter.current === 0) {
+      setIsDragOver(false);
+    }
+  };
+
+  const handleDrop = (e: React.DragEvent) => {
+    e.preventDefault();
+    dragCounter.current = 0;
+    setIsDragOver(false);
+    const files = Array.from(e.dataTransfer.files);
+    if (files.length > 0) {
+      handleFileSelect(files[0]);
+    }
+  };
+
+  const handleFileInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const files = e.target.files;
+    if (files && files.length > 0) {
+      handleFileSelect(files[0]);
+    }
+  };
+
+  const handleBrowseFiles = () => {
+    fileInputRef.current?.click();
+  };
+
+  const handleRemoveFile = () => {
+    setSelectedFile(null);
+    if (fileInputRef.current) {
+      fileInputRef.current.value = "";
+    }
+  };
+
+  const readFileContent = (file: File): Promise<string> => {
+    return new Promise((resolve, reject) => {
+      const reader = new FileReader();
+      reader.onload = () => {
+        const result = reader.result?.toString();
+        if (result) {
+          resolve(result);
+        } else {
+          reject(new Error("Failed to read the file content."));
+        }
+      };
+      reader.onerror = () => {
+        reject(new Error("Error reading file."));
+      };
+      reader.readAsText(file);
+    });
+  };
+
+  return {
+    selectedFile,
+    isDragOver,
+    fileInputRef,
+    isValidFile,
+    handleFileSelect,
+    handleDragOver,
+    handleDragEnter,
+    handleDragLeave,
+    handleDrop,
+    handleFileInputChange,
+    handleBrowseFiles,
+    handleRemoveFile,
+    readFileContent,
+  };
+};
